remove common component
